refactor: 노선에 등록된 역 삭제 불가하도록 수정

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -5,6 +5,9 @@ import org.springframework.stereotype.Service;
 
 import wooteco.subway.exception.DuplicatedStationException;
 import wooteco.subway.exception.NoSuchStationException;
+import wooteco.subway.exception.StationAlreadyInLineException;
+import wooteco.subway.line.dao.LineDao;
+import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
@@ -16,9 +19,11 @@ import java.util.stream.Collectors;
 @Service
 public class StationService {
     private StationDao stationDao;
+    private SectionDao sectionDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public StationResponse saveStation(StationRequest stationRequest) {
@@ -47,6 +52,9 @@ public class StationService {
     }
 
     public void deleteStationById(Long id) {
+        if(sectionDao.findSectionByStationId(id) != 0) {
+            throw new StationAlreadyInLineException();
+        }
         if(stationDao.deleteById(id) == 0) {
             throw new NoSuchStationException();
         }
